Created a signal processing system using the Blinker library
This allows for the pipeline to interact with any interface linked to it. Tested implimentation by creating a highlighting system that shows the node currently running, as well as which nodes have been/have yet to be ran.

The SignalHandler class can be extended to add as many signals as necessary for the pipeline.